warning fixes and preventing crash with JITC fonts.

+++ StarEngine/jni/Graphics/UI/UIStaticTextButton.h
@@ -26,7 +26,7 @@ namespace star
 			const tstring & text,
 			const tstring & fontName,
 			const tstring & fontPath,
-			float32 fontSize,
+			uint32 fontSize,
 			const Color & color,
 			uint32 widthSegements = 1,
 			uint32 heightSegements = 1
